+++ aom-dao/src/main/java/com/github/vaatech/aom/core/repository/MakerRepository.java
@@ -1,15 +1,13 @@
 package com.github.vaatech.aom.core.repository;
 
 import com.github.vaatech.aom.core.model.vehicle.Maker;
+import java.util.List;
+import java.util.Optional;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
-import java.util.List;
-import java.util.Optional;
-
 @Repository
-public interface MakerRepository extends ApplicationJpaRepository<Maker, Long>
-{
+public interface MakerRepository extends ApplicationJpaRepository<Maker, Integer> {
   @Query(
       "SELECT DISTINCT "
           + "m FROM Maker m "
@@ -18,12 +16,10 @@ public interface MakerRepository extends ApplicationJpaRepository<Maker, Long>
   Optional<Maker> findByName(String name);
 
   @Query("SELECT DISTINCT m FROM Maker m LEFT JOIN FETCH m.models mo WHERE m.id = :id")
-  Optional<Maker> findMakerById(Long id);
+  Optional<Maker> findMakerById(Integer id);
 
   @Query("SELECT DISTINCT m FROM Maker m LEFT JOIN FETCH m.models mo")
   List<Maker> findAllWithModels();
 
   boolean existsByName(String name);
-
-
 }
